+++ src/ya_tagscript/adapters/discord_adapters/member_adapter.py
@@ -22,8 +22,11 @@ class MemberAdapter(AttributeAdapter):
       (depends on their top role) (alias: ``colour``)
     - ``colour``: :class:`discord.Colour` — The colour the user's name is shown in
       (depends on their top role) (alias: ``color``)
-    - ``global_name``: :class:`str` | :data:`None` — The user's global nickname
-    - ``nick``: :class:`str` | :data:`None` — The user's guild-specific nickname
+    - ``global_name``: :class:`str` — The user's global nickname (falls back to
+      ``name`` if the user has no global nickname set)
+    - ``nick``: :class:`str` — The user's guild-specific nickname (falls back to
+      ``global_name`` if no guild-specific nickname is set, or to ``name`` if no global
+      nickname is set either)
     - ``avatar`` :class:`tuple[str, Literal[False]]` — The user's avatar. The first
       tuple element contains the avatar's URL. The False instructs the adapter to not
       escape the contents of this attribute.
@@ -57,18 +60,32 @@ class MemberAdapter(AttributeAdapter):
     .. versionchanged:: 1.2
         Now supports passing a :class:`discord.User` as well, with fallback values as
         described above.
+
+    .. versionchanged:: 1.3
+
+        - For members without nicknames, ``nick`` now falls back to ``global_name`` if
+          possible, or ``name`` if the member also has no global nickname.
+        - For users without global nicknames, ``global_name`` now falls back to
+          ``name``.
     """
 
     def __init__(self, member: discord.Member | discord.User):
         super().__init__(base=member)
+
+        # same logic regardless of type
+        global_name = (
+            member.global_name if member.global_name is not None else member.name
+        )
         if isinstance(member, discord.Member):
             joined_at = member.joined_at or member.created_at
-            nick = member.nick
+            # nick if defined, else `global_name` if defined, else `name`
+            nick = member.nick if member.nick is not None else global_name
             top_role = member.top_role
             roles = member.roles
         else:
             joined_at = member.created_at
-            nick = member.global_name
+            # `global_name` if defined, else `name`
+            nick = global_name
             # cannot be None as that is interpreted as "block rejected" when returned
             # by the accessing variable getter block
             top_role = ""
@@ -77,7 +94,7 @@ class MemberAdapter(AttributeAdapter):
         additional_attributes = {
             "color": member.color,
             "colour": member.colour,
-            "global_name": member.global_name,
+            "global_name": global_name,
             "nick": nick,
             "avatar": (member.display_avatar.url, False),
             "discriminator": member.discriminator,